Make Authorized Requests

+++ frontend/src/components/MenuDetails.js
@@ -1,10 +1,18 @@
 import { useMenusContext} from '../hooks/useMenusContext'
+import { useAuthContext } from '../hooks/useAuthContext'
 
 const MenuDetails = ({menu}) => {
     const { dispatch } = useMenusContext()
+    const { user } = useAuthContext()
     const handleClick = async () => {
+        if (!user) {
+            return
+        }
         const response = await fetch('/api/menus/' + menu._id, {
-            method: 'DELETE'
+            method: 'DELETE',
+            headers: {
+                'Authorization': `Bearer ${user.token}`
+            }
         })
         const json = await response.json()
         if (response.ok) {
